+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Rule application logic changed such that a flagged rule continues to run rules in case
   there is a rule that markes it as NoViolation. If I diff is determined to be NoViolation,
   no further rules are run.
+- Fixed issue where the tool sometimes would fail if paths used forward or backslashes.
 
 ## 0.1.7 (2024-11-25)